move FolderFileInfo and FilesContext to avoid imports from app.tsx
Fix imports to avoid circular import and use `import type`
when importing types.
