+++ src/dialogs/rename.tsx
@@ -32,8 +32,8 @@ import { InlineNotification } from 'cockpit-components-inline-notification';
 import type { Dialogs, DialogResult } from 'dialogs';
 import { fmt_to_fragments } from 'utils';
 
-import { FolderFileInfo, useFilesContext } from '../app.tsx';
-import { checkFilename } from '../common.ts';
+import { checkFilename, useFilesContext } from '../common.ts';
+import type { FolderFileInfo } from '../common.ts';
 
 const _ = cockpit.gettext;
