+++ src/download.tsx
@@ -19,7 +19,7 @@
 
 import cockpit from 'cockpit';
 
-import type { FolderFileInfo } from './app.ts';
+import type { FolderFileInfo } from './common.ts';
 
 export function downloadFile(currentPath: string, selected: FolderFileInfo) {
     const payload = JSON.stringify({
