+++ src/upload-button.tsx
@@ -37,8 +37,8 @@ import { superuser } from "superuser";
 import * as timeformat from "timeformat";
 import { fmt_to_fragments } from "utils";
 
-import { FolderFileInfo, useFilesContext } from "./app.tsx";
-import { permissionShortStr } from "./common.ts";
+import { permissionShortStr, useFilesContext } from "./common.ts";
+import type { FolderFileInfo } from "./common.ts";
 import { edit_permissions } from "./dialogs/permissions.tsx";
 import { UploadContext } from "./files-folder-view.tsx";
 import { get_owner_candidates } from "./ownership.tsx";
